refactor: remove old plugin infra
