+++ packages/maskbook/src/background-service.ts
@@ -10,5 +10,4 @@ import './extension/service' // setup Services.*
 import './utils/native-rpc' // setup Android and iOS API server
 import './social-network-adaptor' // setup social network providers
 import './extension/background-script/Jobs' // start jobs
-import './plugins/PluginSerivce' // setup plugins
 import './utils/debug/general'
