+++ packages/maskbook/src/components/InjectedComponents/DecryptedPost/DecryptedPostSuccess.tsx
@@ -1,4 +1,4 @@
-import { memo, useRef, useEffect } from 'react'
+import { memo } from 'react'
 import { useI18N } from '../../../utils'
 import { AdditionalContent, AdditionalContentProps } from '../AdditionalPostContent'
 import { useShareMenu } from '../SelectPeopleDialog'
@@ -7,12 +7,8 @@ import { Link } from '@material-ui/core'
 import type { Profile } from '../../../database'
 import { useStylesExtends } from '../../custom-ui-helper'
 import type { TypedMessage } from '../../../protocols/typed-message'
-import { PluginUI } from '../../../plugins/PluginUI'
-import type { PluginConfig } from '../../../plugins/types'
-import { usePostInfo } from '../../DataSource/usePostInfo'
 import type { ProfileIdentifier } from '../../../database/type'
 import { wrapAuthorDifferentMessage } from './authorDifferentMessage'
-import { ErrorBoundary } from '../../shared/ErrorBoundary'
 import { createInjectHooksRenderer, useActivatedPluginsSNSAdaptor } from '@masknet/plugin-infra'
 
 const PluginRenderer = createInjectHooksRenderer(useActivatedPluginsSNSAdaptor, (x) => x.DecryptedInspector)
@@ -61,7 +57,7 @@ export const DecryptPostSuccess = memo(function DecryptPostSuccess(props: Decryp
         <>
             {shareMenu.ShareMenu}
             <AdditionalContent
-                metadataRenderer={{ after: SuccessDecryptionPlugin }}
+                metadataRenderer={{ after: PluginRenderer }}
                 headerActions={wrapAuthorDifferentMessage(author, postedBy, rightActions)}
                 title={t('decrypted_postbox_title')}
                 message={content}
@@ -70,33 +66,3 @@ export const DecryptPostSuccess = memo(function DecryptPostSuccess(props: Decryp
         </>
     )
 })
-function SuccessDecryptionPlugin(props: PluginSuccessDecryptionComponentProps) {
-    return (
-        <>
-            <PluginRenderer message={props.message} />
-            {[...PluginUI.values()].map((x) => (
-                <ErrorBoundary subject={`Plugin "${x.pluginName}"`} key={x.identifier}>
-                    <OldPluginSuccessDecryptionPostInspectorForEach pluginConfig={x} {...props} />
-                </ErrorBoundary>
-            ))}
-        </>
-    )
-}
-
-function OldPluginSuccessDecryptionPostInspectorForEach(props: { pluginConfig: PluginConfig; message: TypedMessage }) {
-    const { pluginConfig, message } = props
-    const ref = useRef<HTMLDivElement | null>(null)
-    const F = pluginConfig.successDecryptionInspector
-    const post = usePostInfo()
-    useEffect(() => {
-        if (!ref.current || !F || typeof F === 'function') return
-        return F.init(post!, { message }, ref.current)
-    }, [F, post, message])
-    if (!F) return null
-    if (typeof F === 'function') return <F {...post} message={message} />
-    return <div ref={ref} />
-}
-
-interface PluginSuccessDecryptionComponentProps {
-    message: TypedMessage
-}
