+++ packages/maskbook/src/components/InjectedComponents/PageInspector.tsx
@@ -5,10 +5,7 @@ import Close from '@material-ui/icons/Close'
 import IconButton from '@material-ui/core/IconButton'
 import { createInjectHooksRenderer, useActivatedPluginsSNSAdaptor } from '@masknet/plugin-infra'
 import { useMatchXS, MaskMessage, useI18N } from '../../utils'
-import type { PluginConfig } from '../../plugins/types'
-import { PluginUI } from '../../plugins/PluginUI'
 import { useAutoPasteFailedDialog } from './AutoPasteFailedDialog'
-import { ErrorBoundary } from '../shared/ErrorBoundary'
 
 const PluginRender = createInjectHooksRenderer(useActivatedPluginsSNSAdaptor, (x) => x.GlobalInjection)
 export interface PageInspectorProps {}
@@ -51,17 +48,6 @@ export function PageInspector(props: PageInspectorProps) {
         <>
             {JSX}
             <PluginRender />
-            {[...PluginUI.values()].map((x) => (
-                <ErrorBoundary subject={`Plugin "${x.pluginName}"`} key={x.identifier}>
-                    <OldPluginPageInspectorForEach config={x} />
-                </ErrorBoundary>
-            ))}
         </>
     )
-}
-
-function OldPluginPageInspectorForEach({ config }: { config: PluginConfig }) {
-    const F = config.PageComponent
-    if (typeof F === 'function') return <F />
-    return null
 }
