+++ packages/maskbook/src/extension/options-page/DashboardRouters/Plugins.tsx
@@ -1,11 +1,8 @@
-import { useState } from 'react'
-import { makeStyles, ThemeProvider } from '@material-ui/core/styles'
-import { useI18N, extendsTheme } from '../../../utils'
+import { makeStyles } from '@material-ui/core/styles'
+import { useI18N } from '../../../utils'
 import PluginCard from '../DashboardComponents/PluginCard'
 
 import DashboardRouterContainer from './Container'
-import { PluginUI } from '../../../plugins/PluginUI'
-import { PluginScope } from '../../../plugins/types'
 import { useActivatedPluginsSNSAdaptor } from '@masknet/plugin-infra'
 
 const useStyles = makeStyles((theme) => ({
@@ -29,45 +26,26 @@ const useStyles = makeStyles((theme) => ({
     },
 }))
 
-const pluginsTheme = extendsTheme((theme) => ({}))
-
 export default function DashboardSettingsRouter() {
     const { t } = useI18N()
     const classes = useStyles()
-    const [search, setSearch] = useState('')
-    const [searchUI, setSearchUI] = useState('')
     const plugins = useActivatedPluginsSNSAdaptor()
 
     return (
         <DashboardRouterContainer title={t('plugins')}>
-            <ThemeProvider theme={pluginsTheme}>
-                <ul className={classes.pluginList}>
-                    {[...PluginUI.values()]
-                        .filter((plugin) => plugin.scope === PluginScope.Public)
-                        .map((plugin) => (
-                            <li className={classes.pluginItem} key={plugin.id}>
-                                <PluginCard
-                                    key={plugin.id}
-                                    name={plugin.pluginName}
-                                    icon={plugin.pluginIcon}
-                                    id={plugin.id}
-                                    description={plugin.pluginDescription}
-                                />
-                            </li>
-                        ))}
-                    {plugins.map((plugin) => (
-                        <li className={classes.pluginItem} key={plugin.ID}>
-                            <PluginCard
-                                key={plugin.ID}
-                                name={plugin.name.fallback}
-                                id={plugin.ID}
-                                icon={plugin.icon}
-                                description={plugin.description?.fallback}
-                            />
-                        </li>
-                    ))}
-                </ul>
-            </ThemeProvider>
+            <ul className={classes.pluginList}>
+                {plugins.map((plugin) => (
+                    <li className={classes.pluginItem} key={plugin.ID}>
+                        <PluginCard
+                            key={plugin.ID}
+                            name={plugin.name.fallback}
+                            id={plugin.ID}
+                            icon={plugin.icon}
+                            description={plugin.description?.fallback}
+                        />
+                    </li>
+                ))}
+            </ul>
         </DashboardRouterContainer>
     )
 }
