+++ packages/maskbook/src/extension/options-page/index.tsx
@@ -4,7 +4,7 @@ import '../../setup.ui'
 import { useState } from 'react'
 import { useAsync } from 'react-use'
 import { CssBaseline, NoSsr, CircularProgress, Box, Typography, Card } from '@material-ui/core'
-import { ThemeProvider, makeStyles } from '@material-ui/core/styles'
+import { makeStyles } from '@material-ui/core/styles'
 
 import PeopleOutlinedIcon from '@material-ui/icons/PeopleOutlined'
 import CreditCardIcon from '@material-ui/icons/CreditCard'
@@ -13,7 +13,7 @@ import SettingsOutlinedIcon from '@material-ui/icons/SettingsOutlined'
 import PowerIcon from '@material-ui/icons/Power'
 import { HashRouter as Router, Route, Switch, Redirect, useHistory } from 'react-router-dom'
 
-import { useI18N, useClassicMaskTheme, SSRRenderer, Flags, useMatchXS } from '../../utils'
+import { useI18N, SSRRenderer, Flags, useMatchXS } from '../../utils'
 
 import FooterLine from './DashboardComponents/FooterLine'
 import Drawer from './DashboardComponents/Drawer'
@@ -33,9 +33,7 @@ import { SetupStep } from './SetupStep'
 import DashboardNavRouter from './DashboardRouters/Nav'
 import ActionButton from './DashboardComponents/ActionButton'
 import ShowcaseBox from './DashboardComponents/ShowcaseBox'
-import type { PluginConfig } from '../../plugins/types'
-import { PluginUI } from '../../plugins/PluginUI'
-import { ErrorBoundary, withErrorBoundary } from '../../components/shared/ErrorBoundary'
+import { withErrorBoundary } from '../../components/shared/ErrorBoundary'
 import { MaskUIRoot } from '../../UIRoot'
 import { createInjectHooksRenderer, startPluginDashboard, useActivatedPluginsDashboard } from '@masknet/plugin-infra'
 import { createPluginHost } from '../../plugin-infra/host'
@@ -207,25 +205,6 @@ function DashboardUI() {
     )
 }
 
-//#region dashboard plugin UI
-function PluginDashboardInspectorForEach({ config }: { config: PluginConfig }) {
-    const F = config.DashboardComponent
-    if (typeof F === 'function') return <F />
-    return null
-}
-
-function OldPluginRender() {
-    return (
-        <ThemeProvider theme={useClassicMaskTheme()}>
-            {[...PluginUI.values()].map((x) => (
-                <ErrorBoundary key={x.identifier} subject={`Plugin "${x.pluginName}"`}>
-                    <PluginDashboardInspectorForEach config={x} />
-                </ErrorBoundary>
-            ))}
-        </ThemeProvider>
-    )
-}
-//#endregion
 const PluginRender = createInjectHooksRenderer(useActivatedPluginsDashboard, (x) => x.GlobalInjection)
 
 export function Dashboard() {
@@ -236,7 +215,6 @@ export function Dashboard() {
                     <CssBaseline />
                     <DashboardUI />
                     <PluginRender />
-                    <OldPluginRender />
                 </Router>
             </NoSsr>
         </DashboardSnackbarProvider>,
