+++ packages/maskbook/src/plugins/Collectible/SNSAdaptor/Collectible.tsx
@@ -28,7 +28,7 @@ import { HistoryTab } from './HistoryTab'
 import { CollectibleState } from '../hooks/useCollectibleState'
 import { CollectibleCard } from './CollectibleCard'
 import { CollectibleProviderIcon } from './CollectibleProviderIcon'
-import { PluginSkeleton } from '../../PluginSkeleton'
+import { PluginSkeleton } from './PluginSkeleton'
 import { CollectibleProvider, CollectibleTab } from '../types'
 import { currentCollectibleProviderSettings } from '../settings'
 import { MaskbookTextIcon } from '../../../resources/MaskbookIcon'
