+++ packages/maskbook/src/plugins/Collectible/SNSAdaptor/PluginSkeleton.tsx
